Local Cameras
Finalize "standalone" scenario with support for "local cameras". A local camera is one available locally to a portal instance.

With these changes, local cameras can be viewed and modified using the same UI and API as remote cameras. This makes it possible to have a fully-functional lunacam system on a single device.